More work on *Any relationships.
Redo the m2e plugin exclusions to accomidate new tools.
